config/requirements.txt
Django==5.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0
python-decouple==3.8
Pillow==10.2.0
django-cors-headers